Add err handling for the extension with the corresponding error codes.

+++ csm-extensions/services/mysql/CSM_HOME/connection/delete/deleteConnections.sh
@@ -13,7 +13,10 @@ NewUsername=`echo ${CREDENTIALS_ID} |  cut -c 1-15`
 write_success_output () { 
 	cat <<EOF > ${OUTPUT_FILE}
 {
-	"status": "successful"
+	"http_code":200,
+	"status": "successful",
+	"details":"user deleted",
+	"processing_type":"Extension"
 }
 EOF
 }
@@ -22,10 +25,20 @@ EOF
 write_failed_output(){
 	cat <<EOF > ${OUTPUT_FILE}
 {
-	"status": "failed"
+	"http_code" : 500, 
+	"details" : "$1", 
+	"status": "failed",
+	"processing_type" : "Extension"
 }
 EOF
 }
+
+mysql -h ${MYSQL_SERVICE_HOST} -P ${MYSQL_SERVICE_PORT_MYSQL} -u ${Username} -p${Password} -e "select * from mysql.user where User=\"d${NewUsername}\"" | grep ${NewUsername} > /dev/null 2>&1
+
+if [ $? -ne 0 ]; then
+	write_failed_output "User not found"
+	exit 0
+fi
 
 # delete user from mysql
 mysql -h ${MYSQL_SERVICE_HOST} -P ${MYSQL_SERVICE_PORT_MYSQL} -u ${Username} -p${Password} -e "DROP USER \"d${NewUsername}\"@\"%\"; FLUSH PRIVILEGES;" > /dev/null 2>&1
@@ -39,4 +52,4 @@ if [ $? -ne 0 ]; then
 	write_success_output
 	exit 0
 fi
-write_failed_output
+write_failed_output "User could not be deleted"
